+++ pkgs/tui/src/domain.ts
@@ -1,8 +1,8 @@
 import type { Option } from '@hazae41/option'
-import type { Asset, Pair, Rate } from '../../../pkgs/xoswap/src/types.js'
+import type { Asset, Pair, Rate } from 'xoswap'
 
 // Re-export types from xoswap for convenience
-export type { Asset, AssetId, Pair, PairId, Rate } from '../../../pkgs/xoswap/src/types.js'
+export type { Asset, AssetId, Pair, PairId, Rate } from 'xoswap'
 
 // TUI-specific domain types
 export interface SwapQuote {
