+++ pkgs/tui/src/state/swapActions.ts
