+++ pkgs/tui/src/ui/AssetsTable.tsx
